+++ domain/dtos/record_dto.py
@@ -1,10 +1,13 @@
 from data.model.event_type import EventType
 from pydantic import BaseModel
 
+from domain.dtos.blob_stats_dto import BlobStatsDto
+from domain.dtos.league_dto import LeagueDto
+
 
 class RecordDto(BaseModel):
     id: int
-    league_id: int
+    league: LeagueDto
     event_type: EventType
-    competitor_id: int
+    blob: BlobStatsDto
     score: float