+++ web-ui/src/components/pages/HallOfFamePage.tsx
@@ -40,7 +40,7 @@ export const HallOfFamePage = () => {
         <TitlesTab setLoading={setLoading} />
       </Box>
       <Box role="tabpanel" hidden={index !== 2} id="hall-of-fame-tab-events-content">
-        <EventsTab />
+        <EventsTab setLoading={setLoading} />
       </Box>
     </PageFrame>
   );